+++ runtime/test/schnorr.test.ts
@@ -1,0 +1,44 @@
+// This file is part of Compact.
+// Copyright (C) 2025 Midnight Foundation
+// SPDX-License-Identifier: Apache-2.0
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { describe, expect, test } from 'vitest';
+import * as runtime from '../src/index.js';
+
+const MSG_LEN = 3;
+const msgType = new runtime.CompactTypeVector(MSG_LEN, runtime.CompactTypeField);
+const sampleMsg = (): bigint[] => [1n, 2n, 3n];
+
+describe('JubJub Schnorr (TypeScript)', () => {
+  test('sign/verify roundtrip', () => {
+    const sk = runtime.jubjubSampleScalar();
+    const pk = runtime.jubjubSchnorrVerifyingKey(sk);
+    const msg = sampleMsg();
+
+    const sig = runtime.jubjubSchnorrSign(msgType, msg, sk);
+
+    expect(runtime.jubjubSchnorrVerify(msgType, msg, pk, sig)).toBe(true);
+  });
+
+  test('verify rejects a bad signature', () => {
+    const sk = runtime.jubjubSampleScalar();
+    const pk = runtime.jubjubSchnorrVerifyingKey(sk);
+    const msg = sampleMsg();
+
+    const sig = runtime.jubjubSchnorrSign(msgType, msg, sk);
+    const badSig = { ...sig, response: (sig.response + 1n) % runtime.JUBJUB_SCALAR_ORDER };
+
+    expect(runtime.jubjubSchnorrVerify(msgType, msg, pk, badSig)).toBe(false);
+  });
+});
